requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
django-filter==23.2
pytest==7.4.2
pytest-django==4.5.2


